Fix size calculation for writeTexture
I've run into this a bunch because it's confusing but

size (the 4th parameter to writeTexture) is in pixels
but rowsPerImage is in blocks.

Further, while each row of data has to be aligned to bytesPerRow,
the last row does not need to be bytesPerRow long. It only needs
however many bytes will actually be read in that row.

To fix this requires knowing the size of a block in pixels and
the number of bytes per block so extracted that data from the CTS